fix(#259): move rating window logic to rating service
