+++ src/backend/rating_app/services/__init__.py
@@ -3,7 +3,6 @@ from .course_service import CourseService
 from .department_service import DepartmentService
 from .faculty_service import FacultyService
 from .instructor_service import InstructorService
-from .rating_policy import RatingWindowPolicy
 from .rating_service import RatingService
 from .semester_service import SemesterService
 from .speciality_service import SpecialityService
@@ -12,7 +11,6 @@ from .student_service import StudentService
 __all__ = [
     "CourseService",
     "RatingService",
-    "RatingWindowPolicy",
     "InstructorService",
     "StudentService",
     "FacultyService",
